requirements.txt
pyside6>=6.5.0
pandas>=2.0.0
openpyxl>=3.1.0
sqlalchemy>=2.0.0
reportlab>=4.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
lxml>=4.9.0
pyarrow>=10.0.0
tables>=3.8.0
pyyaml>=6.0.0
pytest>=7.0.0
pytest-cov>=4.0.0